pyproject.toml
[tool.poetry]
name = "controller-new"
version = "0.1.0"
description = ""
authors = ["Mark <74201924+Markushik@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "controller_new"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.18"
nats-py = "^2.3.1"
fluent-runtime = "^0.4.0"
dynaconf = { extras = ["redis"], version = "^3.1.12" }
ormsgpack = "^1.2.6"
asyncpg = "^0.28.0"
alembic = "^1.11.1"
pathvalidate = "^3.0.0"
markupsafe = "^2.1.3"
taskiq-nats = "^0.3.0"
uuid6 = "^2023.5.2"
adaptix = "^3.0.0a4"
orjson = "^3.9.5"
aiogram-dialog = "^2.0.0"
zstd = "^1.5.5.1"
structlog = "^23.1.0"
colorama = "^0.4.6"
loguru = "^0.7.2"
taskiq = {extras = ["reload"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.280"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
